AML Competition Changes - WE ADVANCED TO REGIONALS!! - Refinements to autonomous
Throughout the competition, small changes were made to directions and distances traveled to be more consistent on the meet's practice and match fields.

Allowed for autonomous to be more consistent

Thoughts going into next comp:
Implement a directional/positional PID to ensure that inconsistencies in the chassis of the robot are adjusted for. Constantly changing the directions to accomodate for these inconsistencies was very tedious, stressful, and risky.