Refinement: Red and Blue Autos
Added transport for third Skystone and platform moving and parking on both the red and blue side

For blue side, the only combination that is not yet fully coded is the right skystone orientation for the blue side. Currently transports the first Skystone and then moves to the second Skystone.